refactor: Extract logic from SupplementController
Refactored `SupplementController::index` to use a dedicated Action class `FetchSupplementsIndexAction`.
Moved `getSupplementsWithLatestLog`, `getUsageHistory`, and `fillUsageHistory` logic to the new Action class.
Cleaned up controller imports and methods.
Verified with feature tests.
